back-end/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
resend==0.7.0
python-decouple==3.8